renamed signpost energy to signpost_energy_monitors
