+++ apps/controller/enable_i2c/main.c
@@ -9,7 +9,6 @@
 #include <tock.h>
 
 #include "controller.h"
-#include "signpost_energy.h"
