+++ apps/storage_master/api_test/main.c
@@ -13,7 +13,6 @@
 
 #include "app_watchdog.h"
 #include "signpost_api.h"
-#include "signpost_energy.h"
 #include "signpost_storage.h"
 #include "storage_master.h"
