+++ apps/storage_master/signpost_storage_master_app/main.c
@@ -14,7 +14,6 @@
 
 #include "app_watchdog.h"
 #include "signpost_api.h"
-#include "signpost_energy.h"
 #include "signpost_storage.h"
 #include "storage_master.h"
